--- packages/frontend/core/src/modules/workspace-indexer-embedding/entities/embedding.ts
@@ -1,290 +1,0 @@
-import type { WorkspaceService } from '@affine/core/modules/workspace';
-import { DebugLogger } from '@affine/debug';
-import type { PaginationInput } from '@affine/graphql';
-import {
-  catchErrorInto,
-  effect,
-  Entity,
-  fromPromise,
-  LiveData,
-  onComplete,
-  onStart,
-  smartRetry,
-} from '@toeverything/infra';
-import { EMPTY, interval, Subject } from 'rxjs';
-import { exhaustMap, mergeMap, switchMap, takeUntil } from 'rxjs/operators';
-
-import { COUNT_PER_PAGE } from '../constants';
-import type { EmbeddingStore } from '../stores/embedding';
-import type {
-  IgnoredDoc,
-  LocalAttachmentFile,
-  PersistedAttachmentFile,
-} from '../types';
-
-const logger = new DebugLogger('WorkspaceEmbedding');
-
-export interface EmbeddingConfig {
-  enabled: boolean;
-}
-
-interface Attachments {
-  totalCount: number;
-  pageInfo: {
-    endCursor: string | null;
-    hasNextPage: boolean;
-  };
-  edges: {
-    node: PersistedAttachmentFile;
-  }[];
-}
-
-type IgnoredDocs = IgnoredDoc[];
-
-interface EmbeddingProgress {
-  embedded: number;
-  total: number;
-}
-
-export class Embedding extends Entity {
-  enabled$ = new LiveData<boolean | null>(null);
-  error$ = new LiveData<any>(null);
-  attachments$ = new LiveData<Attachments>({
-    edges: [],
-    pageInfo: {
-      endCursor: null,
-      hasNextPage: false,
-    },
-    totalCount: 0,
-  });
-  ignoredDocs$ = new LiveData<IgnoredDocs>([]);
-  isEnabledLoading$ = new LiveData(true);
-  isAttachmentsLoading$ = new LiveData(true);
-  isIgnoredDocsLoading$ = new LiveData(true);
-  embeddingProgress$ = new LiveData<EmbeddingProgress | null>(null);
-  isEmbeddingProgressLoading$ = new LiveData(true);
-
-  private readonly EMBEDDING_PROGRESS_POLL_INTERVAL = 3000;
-  private readonly stopEmbeddingProgress$ = new Subject<void>();
-  uploadingAttachments$ = new LiveData<LocalAttachmentFile[]>([]);
-
-  constructor(
-    private readonly workspaceService: WorkspaceService,
-    private readonly store: EmbeddingStore
-  ) {
-    super();
-  }
-
-  mergedAttachments$ = LiveData.computed(get => {
-    const uploading = get(this.uploadingAttachments$);
-    const uploaded = get(this.attachments$).edges.map(edge => edge.node);
-    return [...uploading, ...uploaded].slice(0, 10);
-  });
-
-  getEnabled = effect(
-    exhaustMap(() => {
-      return fromPromise(signal =>
-        this.store.getEnabled(this.workspaceService.workspace.id, signal)
-      ).pipe(
-        smartRetry(),
-        mergeMap(value => {
-          this.enabled$.next(value);
-          return EMPTY;
-        }),
-        catchErrorInto(this.error$, error => {
-          logger.error(
-            'Failed to fetch workspace doc embedding enabled',
-            error
-          );
-        }),
-        onStart(() => this.isEnabledLoading$.setValue(true)),
-        onComplete(() => this.isEnabledLoading$.setValue(false))
-      );
-    })
-  );
-
-  setEnabled = (enabled: boolean) => {
-    return this.store
-      .updateEnabled(this.workspaceService.workspace.id, enabled)
-      .then(() => {
-        this.getEnabled();
-      });
-  };
-
-  getIgnoredDocs = effect(
-    exhaustMap(() => {
-      return fromPromise(signal =>
-        this.store.getIgnoredDocs(this.workspaceService.workspace.id, signal)
-      ).pipe(
-        smartRetry(),
-        mergeMap(value => {
-          this.ignoredDocs$.next(value);
-          return EMPTY;
-        }),
-        catchErrorInto(this.error$, error => {
-          logger.error(
-            'Failed to fetch workspace doc embedding ignored docs',
-            error
-          );
-        }),
-        onStart(() => this.isIgnoredDocsLoading$.setValue(true)),
-        onComplete(() => this.isIgnoredDocsLoading$.setValue(false))
-      );
-    })
-  );
-
-  updateIgnoredDocs = ({
-    add,
-    remove,
-  }: {
-    add: string[];
-    remove: string[];
-  }) => {
-    return this.store
-      .updateIgnoredDocs(this.workspaceService.workspace.id, add, remove)
-      .then(() => {
-        this.getIgnoredDocs();
-      });
-  };
-
-  getAttachments = effect(
-    exhaustMap((pagination: PaginationInput) => {
-      return fromPromise(signal =>
-        this.store.getEmbeddingFiles(
-          this.workspaceService.workspace.id,
-          pagination,
-          signal
-        )
-      ).pipe(
-        smartRetry(),
-        mergeMap(value => {
-          const patched = {
-            ...value,
-            edges: value.edges.map(edge => ({
-              ...edge,
-              node: {
-                ...edge.node,
-                status: 'uploaded' as const,
-              },
-            })),
-          };
-          this.attachments$.next(patched);
-          return EMPTY;
-        }),
-        catchErrorInto(this.error$, error => {
-          logger.error(
-            'Failed to fetch workspace doc embedding attachments',
-            error
-          );
-        }),
-        onStart(() => this.isAttachmentsLoading$.setValue(true)),
-        onComplete(() => this.isAttachmentsLoading$.setValue(false))
-      );
-    })
-  );
-
-  addAttachments = (files: File[]) => {
-    const generateLocalId = () =>
-      Math.random().toString(36).slice(2) + Date.now();
-    const localAttachments: LocalAttachmentFile[] = files.map(file => ({
-      localId: generateLocalId(),
-      fileName: file.name,
-      mimeType: file.type,
-      size: file.size,
-      createdAt: file.lastModified,
-      status: 'uploading',
-    }));
-
-    this.uploadingAttachments$.next([
-      ...localAttachments,
-      ...this.uploadingAttachments$.value,
-    ]);
-
-    this.store
-      .addEmbeddingFiles(this.workspaceService.workspace.id, files)
-      .then(() => {
-        this.uploadingAttachments$.next(
-          this.uploadingAttachments$.value.filter(
-            att => !localAttachments.some(l => l.localId === att.localId)
-          )
-        );
-        this.getAttachments({ first: COUNT_PER_PAGE, after: null });
-      })
-      .catch(error => {
-        this.uploadingAttachments$.next(
-          this.uploadingAttachments$.value.map(att =>
-            localAttachments.some(l => l.localId === att.localId)
-              ? { ...att, status: 'error', errorMessage: String(error) }
-              : att
-          )
-        );
-      });
-  };
-
-  removeAttachment = (id: string) => {
-    const localIndex = this.uploadingAttachments$.value.findIndex(
-      att => att.localId === id
-    );
-    if (localIndex !== -1) {
-      this.uploadingAttachments$.next(
-        this.uploadingAttachments$.value.filter(att => att.localId !== id)
-      );
-      return Promise.resolve();
-    }
-    return this.store
-      .removeEmbeddingFile(this.workspaceService.workspace.id, id)
-      .then(() => {
-        this.getAttachments({ first: COUNT_PER_PAGE, after: null });
-      });
-  };
-
-  startEmbeddingProgressPolling() {
-    this.stopEmbeddingProgressPolling();
-    this.getEmbeddingProgress();
-  }
-
-  stopEmbeddingProgressPolling() {
-    this.stopEmbeddingProgress$.next();
-  }
-
-  getEmbeddingProgress = effect(
-    exhaustMap(() => {
-      return interval(this.EMBEDDING_PROGRESS_POLL_INTERVAL).pipe(
-        takeUntil(this.stopEmbeddingProgress$),
-        switchMap(() =>
-          fromPromise(signal =>
-            this.store.getEmbeddingProgress(
-              this.workspaceService.workspace.id,
-              signal
-            )
-          ).pipe(
-            smartRetry(),
-            mergeMap(value => {
-              this.embeddingProgress$.next(value);
-              if (value && value.embedded === value.total) {
-                this.stopEmbeddingProgressPolling();
-              }
-              return EMPTY;
-            }),
-            catchErrorInto(this.error$, error => {
-              logger.error(
-                'Failed to fetch workspace embedding progress',
-                error
-              );
-            }),
-            onStart(() => this.isEmbeddingProgressLoading$.setValue(true)),
-            onComplete(() => this.isEmbeddingProgressLoading$.setValue(false))
-          )
-        )
-      );
-    })
-  );
-
-  override dispose(): void {
-    this.getEnabled.unsubscribe();
-    this.getAttachments.unsubscribe();
-    this.getIgnoredDocs.unsubscribe();
-    this.stopEmbeddingProgress$.next();
-    this.getEmbeddingProgress.unsubscribe();
-  }
-}
