+++ packages/frontend/core/src/modules/workspace-indexer-embedding/services/embedding.ts
@@ -1,7 +1,13 @@
 import { Service } from '@toeverything/infra';
 
-import { Embedding } from '../entities/embedding';
+import { AdditionalAttachments } from '../entities/additional-attachments';
+import { EmbeddingEnabled } from '../entities/embedding-enabled';
+import { EmbeddingProgress } from '../entities/embedding-progress';
+import { IgnoredDocs } from '../entities/ignored-docs';
 
 export class EmbeddingService extends Service {
-  embedding = this.framework.createEntity(Embedding);
+  embeddingEnabled = this.framework.createEntity(EmbeddingEnabled);
+  additionalAttachments = this.framework.createEntity(AdditionalAttachments);
+  ignoredDocs = this.framework.createEntity(IgnoredDocs);
+  embeddingProgress = this.framework.createEntity(EmbeddingProgress);
 }
